+++ implementacao.txt
@@ -9,21 +9,14 @@ mover coisa repetida para superclasses (abstratas, as vezes
 
 usar doxygen nos arquivos .hpp das classes
 
-tab size de 4 para 2, em todos arquivos
-
-cliar arquivos .cpp para as tres classes
-	- ignorar makefiles se não compilar
-
-separar em pastas se der NÃO pau
+nem tudo precisa ser public
 
 organizar construtores depois
 
-  EncomendaNormal* enc0 = new EncomendaNormal(5, 12, *cli0, *cli1)
-isso pode dar PAU!
+- ver se os makefiles podem dar pau
 
-cria um vetor de encomendas na main para evitar repetição daquele jeito
-
-*** EXLUIR MAIN.EXE e ARQUIVOS TXT antes de dar PR ****
+*** EXLUIR ARQUIVOS TXT antes de dar PR ****
+**** MOVER MAIN.EXE PARA FORA ANTES DO PR ***
 
 ------ IMPLEMENTAÇÃO ----
 1) main.cpp
