+++ includes/Cliente.hpp
@@ -2,9 +2,11 @@
 #define CLIENTE_H
 
 #include <string>
+#include <iostream>
+
 using namespace std;
 
-class Cliente{
+class Cliente {
 
 	public:
 		string _name;
