+++ includes/EncomendaNormal.hpp
@@ -1,11 +1,14 @@
 #ifndef EncomendaNormal_H
 #define EncomendaNormal_H
 
-#include "Encomenda.hpp"
-#include "Cliente.hpp"
+#include <iostream>
+
+#include "./Encomenda.hpp"
+#include "./Cliente.hpp"
+
 using namespace std;
 
-class EncomendaNormal: public Encomenda{
+class EncomendaNormal: public Encomenda {
 
 	public:
